# docs/plans/api-specification.yaml
openapi: 3.1.0
info:
  title: ekaiX AIXcelerator API
  version: 1.0.0
  description: |
    Backend API for ekaiX — a Snowflake Native App that enables business users to create
    semantic models and Cortex Agents through AI conversation.

    **Authentication:** All requests are authenticated via the `Sf-Context-Current-User` header
    injected by SPCS (Snowpark Container Services). No separate auth system exists.

    **Workspace Isolation:** All data is scoped to the authenticated user's workspace via
    PostgreSQL Row-Level Security. Sharing is explicit via the shares endpoints.

    **Streaming:** Agent responses are delivered via Server-Sent Events (SSE) on the
    `/agent/stream/{session_id}` endpoint.
  contact:
    name: ekai Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{spcs_endpoint}
    description: SPCS production
    variables:
      spcs_endpoint:
        default: ekaix.snowflakecomputing.app
...
security:
  - spcsAuth: []
components:
  securitySchemes:
    spcsAuth:
      type: apiKey
      in: header
      name: Sf-Context-Current-User
      description: Snowflake username injected by SPCS. Cannot be spoofed in production.

  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code (e.g., NOT_FOUND, VALIDATION_ERROR, FORBIDDEN)
          examples: ["NOT_FOUND"]
        message:
          type: string
          description: Human-readable error description
          examples: ["Data product not found"]
        details:
          type: object
          description: Additional error context (validation errors, field-level issues)
          additionalProperties: true
    PaginationMeta:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0

    # --- Auth ---
...
    DataProduct:
      type: object
      required: [id, name, database_reference, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          examples: ["Revenue Recognition Engine"]
        description:
          type: string
        database_reference:
          type: string
          examples: ["FINANCE_PROD_DB"]
        schemas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [discovery, requirements, generation, validation, published, archived]
        health_score:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        current_phase:
          type: string
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        published_agent_fqn:
          type: string
          nullable: true
        owner:
          type: string
        share_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DataProductCreate:
      type: object
      required: [name, database_reference, schemas]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        description:
          type: string
          maxLength: 2000
        database_reference:
          type: string
          minLength: 1
        schemas:
          type: array
          items:
            type: string
          minItems: 1
    DataProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        description:
          type: string
          maxLength: 2000
        status:
          type: string
          enum: [discovery, requirements, generation, validation, published, archived]
    ShareRequest:
      type: object
      required: [shared_with_user, permission]
      properties:
        shared_with_user:
          type: string
          description: Snowflake username to share with
        permission:
          type: string
          enum: [view, edit]
    ShareInfo:
      type: object
      required: [id, shared_with_user, permission, shared_by, created_at]
      properties:
        id:
          type: string
          format: uuid
        shared_with_user:
          type: string
        permission:
          type: string
          enum: [view, edit]
        shared_by:
          type: string
        created_at:
          type: string
          format: date-time

    # --- Agent ---
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchParam:
      name: search
      in: query
      schema:
        type: string
      description: Full-text search filter

# ============================================================================
# PATHS
# ============================================================================
paths:
...
  /data-products:
    get:
      tags: [Data Products]
      operationId: listDataProducts
      summary: List user's data products
      description: Returns data products owned by or shared with the current user.
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - $ref: "#/components/parameters/SearchParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [discovery, requirements, generation, validation, published, archived]
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, updated_at, created_at, status, health_score]
            default: updated_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated list of data products
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/DataProduct"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"

    post:
      tags: [Data Products]
      operationId: createDataProduct
      summary: Create a new data product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataProductCreate"
      responses:
        "201":
          description: Data product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataProduct"
        "400":
          description: Validation error (duplicate name, invalid database)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Data product with this name already exists in workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /data-products/{id}:
    get:
      tags: [Data Products]
      operationId: getDataProduct
      summary: Get data product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Data product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataProduct"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Data Products]
      operationId: updateDataProduct
      summary: Update a data product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataProductUpdate"
      responses:
        "200":
          description: Updated data product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataProduct"
        "403":
          description: No edit permission (view-only share)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Data Products]
      operationId: deleteDataProduct
      summary: Delete a data product
      description: |
        Deletes the data product and all associated artifacts, BRDs, semantic views,
        documents, and audit logs. Only the owner can delete.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted successfully
        "403":
          description: Only the owner can delete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /data-products/{id}/share:
    get:
      tags: [Data Products]
      operationId: listShares
      summary: List shares for a data product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of shares
          content:
            application/json:
              schema:
                type: object
                required: [shares]
                properties:
                  shares:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShareInfo"

    post:
      tags: [Data Products]
      operationId: shareDataProduct
      summary: Share data product with another user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareRequest"
      responses:
        "201":
          description: Share created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareInfo"
        "400":
          description: Cannot share with self or invalid user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already shared with this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /data-products/{id}/share/{share_id}:
    delete:
      tags: [Data Products]
      operationId: revokeShare
      summary: Revoke a share
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: share_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Share revoked
        "403":
          description: Only the owner can revoke shares
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...